VLE/QRW-74/publish functionality

+++ user_profile/views.py
@@ -19,11 +19,12 @@ def profile(request):
         'username': request.user.username,
         'email': request.user.email,
         'image_url': request.user.profile.image.url,
-        'user_pages':user_pages
+        'user_pages': user_pages
     }
-    return render(request,'user_profile/profile.html',context)
+    return render(request, 'user_profile/profile.html', context)
 
 
+@login_required
 def change_password(request):
     """change password view"""
     if request.method == 'POST':
@@ -43,15 +44,16 @@ def change_password(request):
     })
 
 
+@login_required
 def change_image(request):
     """change image view"""
     if request.method == 'POST':
         update_image_form = UpdateImageForm(
-            request.POST,request.FILES,instance=request.user.profile)
+            request.POST, request.FILES, instance=request.user.profile)
 
         if update_image_form.is_valid():
             update_image_form.save()
-            messages.success(request,'Your profile image has been updated!')
+            messages.success(request, 'Your profile image has been updated!')
             return redirect('profile')
 
     else:
@@ -63,6 +65,7 @@ def change_image(request):
     return render(request, 'user_profile/change_image.html', context)
 
 
+@login_required
 def delete_image(request):
     """delete_image view"""
     if request.method == 'POST':
@@ -73,6 +76,7 @@ def delete_image(request):
     return redirect('profile')
 
 
+@login_required
 def create_page(request):
     """create_page view"""
     if request.method == 'POST':
@@ -87,18 +91,22 @@ def create_page(request):
     else:
         form = PageForm()
 
-    return render(request, 'constructor/page_update.html', {'form': form})
+    return render(request, 'constructor/update.html', {'form': form})
 
 
 def view_page(request, page_id):
     """view_page view"""
-    page = Page.objects.get(id=page_id)
-    return render(request, 'constructor/page_view.html', {'post': page})
+    page = Page.objects.get(upid=page_id)
+    return render(request, 'constructor/view.html', {'post': page})
 
 
+@login_required
 def delete_page(request, page_id):
     """delete_page view"""
-    page = get_object_or_404(Page, id=page_id)
+    page = get_object_or_404(Page, upid=page_id)
+    if request.user != page.user:
+        messages.warning(request, 'Your are not authorized to delete this page')
+        return redirect('profile')
     if request.method == 'POST':
         page.delete()
     return redirect('profile')
